backend: sql request and cmap to DTO fixed on subscribed POSTs

+++ back/src/main/java/com/openclassrooms/api/repositories/CommentRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CommentRepository extends JpaRepository<Comment, Integer> {
-    // Ajouter des méthodes de requête personnalisées si nécessaire
+public interface CommentRepository extends JpaRepository<Comment, Long> {
+    // Méthodes spécifiques si nécessaire
 }
